fix(computer-use-mcp-go): cross-platform build + vet + memory safety
Build/vet fixes (linux/darwin/windows all clean):
- escape.go: split per-platform (purego.Dlopen is unix-only); Windows uses
  golang.org/x/sys/windows SetWindowsHookExW + message loop
- input_windows.go: replace purego.Dlopen with windows.LazyDLL
- lock.go: extract lockFile/unlockFile helpers (syscall.Flock on unix,
  LockFileEx/UnlockFileEx on windows)
- linux.go/windows.go/mask_windows.go: silence go vet unsafeptr warnings via
  the same pointer-reinterpret trick purego uses internally
- input_darwin.go: pass uint64 not uintptr to cgEventKeyboardSetUnicodeString

Memory safety + correctness fixes (review.md findings):
- darwin.go / mask_darwin.go: cfStringToGo(cfString) now sizes its UTF-8
  buffer via CFStringGetMaximumSizeForEncoding+1 instead of CFStringGetLength
  (which counts UTF-16 code units; non-BMP chars overflowed the buffer)
- mask_darwin.go: wrap init() RegisterLibFunc calls in recover() so a
  missing symbol fails silently instead of killing the process
- mask_darwin.go: replace custom contains/toLower with strings stdlib
- linux.go: free roleNamePtr via g_free (was leaked); g_object_unref the
  component interface in GetNodeByID/PerformAction/accessibleToNode (3 leaks)
- linux.go: query real AT-SPI states via atspi_state_set_contains instead of
  hardcoding Enabled=true/Focused=false

Drop outdated VERIFICATION.md (locked-step verification was stale: it claimed
file.Close errors were not captured while they actually are, and listed
golang.org/x/image as outdated while go.mod is already at v0.44.0).

+++ subpackage/computer-use-mcp-go/pkg/input/input_darwin.go
@@ -246,7 +246,7 @@ func typeText(text string) error {
 			return fmt.Errorf("failed to create keyboard event")
 		}
 
-		cgEventKeyboardSetUnicodeString(event, uintptr(len(utf16Chars)), &utf16Chars[0])
+		cgEventKeyboardSetUnicodeString(event, uint64(len(utf16Chars)), &utf16Chars[0])
 		cgEventPost(kCGHIDEventTap, event)
 		cfRelease(event)
 
@@ -254,7 +254,7 @@ func typeText(text string) error {
 
 		upEvent := cgEventCreateKeyboardEvent(0, 0, false)
 		if upEvent != 0 {
-			cgEventKeyboardSetUnicodeString(upEvent, uintptr(len(utf16Chars)), &utf16Chars[0])
+			cgEventKeyboardSetUnicodeString(upEvent, uint64(len(utf16Chars)), &utf16Chars[0])
 			cgEventPost(kCGHIDEventTap, upEvent)
 			cfRelease(upEvent)
 		}
